added post and put support api

+++ src/main/java/com/app/absworldxpress/model/OrderModel.java
@@ -21,7 +21,7 @@ public class OrderModel {
     private String customerPhoneNumber;
     private String deliveryAddress;
 
-    private double orderAmount;
+    private Integer orderAmount;
 
     @OneToMany(cascade = {CascadeType.DETACH, CascadeType.REFRESH})
     private List<ProductModel> productModelList;
